Change how the screen objects are pushed/popped
Instead of creating a new object with duplicated data, we create an
object pointer pointing to this data. This allows us to do
pointer-equality comparison with the screen objects from the lua
callback functions. This fixes the script to jump to a canvas showing
the desired window in the display.