[PATCH 2/3] screen: Introduce windows command
This new command is doing nearly the same as the windows
command.
But a string escape can be passed which makes it much
more flexible. The default string escape if no argument
is passed is "%n%f %t " which is intended to rebuild
the same output the windows command would give as good
as possible (slight changes with the flags can happen).

This command is also not be limited in output size
(the windows command is limited to 1024 bytes).

The windowsx command can be queried (-Q command) and
this is its main purpose (be able to query the exact
window list status of an active screen session from
shell).

Signed-off-by: Thomas Renninger <address@hidden>
Signed-off-by: Thomas Renninger <Thomas Renninger" address@hidden>

---------

On Tuesday, February 12, 2013 12:07:54 AM Thomas Renninger wrote:
> This new command is doing nearly the same as the windows
> command.

> The default string escape if no argument
> is passed is "%n%f %t " which is intended to rebuild
> the same output the windows command would give as good
> as possible (slight changes with the flags can happen).
>  int where;
Ooops this is wrong, I still changed this so that no output
change happens (minor changes with flags) when the already
existing command:
"windows" without argument is used.
Above should be done in a devel branch where somewhat
bigger cleanups are allowed including slight output changes.

So currently "windows" command without arguments is the same
as before.

   Thomas

https://lists.gnu.org/archive/html/screen-devel/2013-02/msg00004.html
https://lists.gnu.org/archive/html/screen-devel/2013-02/msg00007.html

Applied manually with small changes