Began refactoring conditional escapes %? and %: to use WinMsgCond abstraction
Note that this retains use of omflag (see below) and qmnumrend; the latter is
kept in tact beacuse that has nothing to do with conditions and should stick
with rendition stuff.

With regards to the flags---this was particularily difficult to understand at
first; I'm still uncertain what `q' and `o' stand for, although I have
determined that `q' is set to true when %? should be true and `o' is merely a
flag used to keep track of whether or not %: has been encountered so that
subsequent %:'s and the closing %? are handled properly. Truth be told, it's so
confusing to work with that I'm going to separate the removal of omflag into
another commit (to follow next) so that the diff is more easily followed.