[SECURITY FIX] wmbc_printf now protects against buffer overflows by dynamically expanding the size of the buffer
C99 introduced *snprintf functions that limit the number of characters written
to a buffer; prior to this change, portions of MakeWinMsgEv (and the functions
it calls) were vulnerable to buffer overflows.

I am uncertain why the #define is required at the moment; this will need to be
discussed, as I'm unsure why vsnprintf is being replaced with xvsnprintf as it
is (gcc -E yields nothing).