Window message buffer pointer increment is now encapsulated in wmbc_*
This will greatly simplify logic going forward since we always know that the
pointer will point to the next byte to be written; previous to this change, the
pointer was incremented at the beginning of the MakeWinMsgEv loop, the character
written into the buffer immediately---possibly modified later on---and the
pointer incremented/decremented throughout.

As an exception, the padding code is largely unchanged; this is a larger beast
that will be handled on its own.