Copied tests/{signature.h,macros.h} from gnulib and added check target to Makefile
signature.h was modified to explitily use gcc's unused attribute; if we want to
work back in its macro abstraction so that it works with other compilers, that's
fine...and probably a good idea.

The object file of the same name as the test (after the test-* prefix) is
automatically linked. Tests are automatically executed by the `check' target
after having been compiled.