Added winmsgcond test case using gnulib-style testing
Since gnulib is actively maintained and is intended to represent a high quality,
portable library for the GNU project, I figured this would be an excellent
starting point. It is simple to use and we can always migrate to a framework
later if we so choose to. I also like its idea of raising SIGABRT to generate a
core dump, which is far more useful than conventional frameworks' solutions in
other languages. Granted, that also means it aborts immediately without running
all the other tests, which could be a pita.