make status position runtime configurable
Status covering left side, especially bottom left is very annoying
because it covers the prompt and e.g. with visual bell I have to wait to
see the contents beind this visual signal. The distributions compile the
default status position. I used to recompile but it is tedious. I would
be much easier just to keep the favourite settings in the configuration
file.

Signed-off-by: Amadeusz Sławiński <amade@asmblr.net>