screen: fix breakage when used inside qemu
when a rootfs is used in qemu with virtio mount,
new files created get the uid of the qemu user, not the uid of the
user inside qemu. this is usually no problem since you have full
access permissions to everything - however, screen did some uid
comparison checks, which only make sense when the screen sockets
are stored in a public place such as /var/run or /tmp.
in our case, they are stored in $HOME, so any screen there belongs
to us and we do not need to compare uids.

Signed-off-by: John Spencer <maillist-screen@barfooze.de>
Signed-off-by: Amadeusz Sławiński <amade@asmblr.net>