screen: Implement dead/zombie window polling
Currently if zombie keys are defined, one needs to
explicitly hit a key to tell screen to try to reconnect
a window. This is rather unfortunte if you for example
have dozens of screens connected to foreign machines
through network connections. Once the network connection
is cut for a while, all windows will enter the dead/zombie
state and one has to go through all windows manually
and hit the zombie resurrect key, once the network got
set up again.

This patch implements auto-reconnecting via zombie_timeout
(in seconds) variable. By default it is set to 0 which
complies to current behavior (no polling is done).

Signed-off-by: Alexander Naumov <posix.ru@gmail.com>
Signed-off-by: Thomas Renninger <trenn@suse.de>