screen: Introduce argument to windows command
A string escape can be passed which makes the windows
command much more flexible. The default string escape
if no argument is passed is "%n%f %t " which is intended
to rebuild the same output the windows command would give
as good as possible (slight changes with the flags can
happen).

This command is also not be limited in output size
(the windows command is limited to 1024 bytes).

The windowsx command can be queried (-Q command) and
this is its main purpose (be able to query the exact
window list status of an active screen session from
shell).

Signed-off-by: Alexander Naumov <alexander_naumov@opensuse.org>
Signed-off-by: Thomas Renninger <trenn@suse.de>