Process: Fix sign-compare errors
DoProcess:
oldlen here is defined as a signed integer, with the lenp coming through
as a size_t (unsigned). Convert the local storage to size_t to match, and
allow the compiler to catch any further signed errors later

suFin:
Another len vs local store of len which hasn't been converted to size_t
In this use case, the local l takes the size of the char arrays in
struct inputsu, which are > 0, so this it should be reasonable
to set l as size_t.

Signed-off-by: Kieran Bingham <kieranbingham@gmail.com>