Process: Pointer arithmetic comparison with unsigned size_t
These two fixes cast the resultant pointer calculation to a
uintptr_t. This type is unsigned, (matching size_t) but specifically
set to store pointers.

I dislike simply casting to remove warnings, but in this instance
I think this may be reasonable. The aim of both of these cases is to
loop for the length of the buffer, so one alternative could be to
extract the calculation. If this is preferred by the developers,
we can drop this patch.

Signed-off-by: Kieran Bingham <kieranbingham@gmail.com>