Ignore logfile's name that begins with the "-" symbol
This fixes API:
To enable logging we use -L option. But in case of
default logfile name (screenlog.0) we will need to
define it anyway. Because screen will try to interpret
next option as a parameter for -L option (which is
logfile name). It will fails ALWAYS, because next
parameter will always start with "-" symbol...
what is not permited for logfile name of course.

For example:
$ screen -L -D -m ./configure
In this case logfile name is screenlog.0, because "-D"
will not be interpreted by screen as a name of logfile.

Signed-off-by: Alexander Naumov <alexander_naumov@opensuse.org>