Add argc parameter to Parse function
We want to call commands with argc, argv, so we need to get argc from
Parse command. While argc is being calculated, it's just being returned,
we could just do argc = Parse(), but because argv is also one of
parameters, it looks better to also make argc into a parameter.

Signed-off-by: Amadeusz Sławiński <amade@asmblr.net>