Use DoCommand instead of DoAction
We trigger kill command after confirming that user wants to kill a
window on windowlist. It was done using DoAction, however with goal of
removing DoAction, migrate to DoCommand. We use Parse, which is probably
bit overkill in this case, but it's how it's done in other places, so
let's be consistent for now and reconsider later.

Signed-off-by: Amadeusz Sławiński <amade@asmblr.net>