Disable exclusive mode on TTY device when closing
When opening a TTY we enable exclusive mode to prevent other tools
messing with our connection. When we are done using the TTY the
exclusive mode must be disabled so the TTY can be reconnected to
later. We remember to do this when a break is sent to the device,
but not on window close.

This change ensures we disable exclusive mode on the TTY device
whenever the window is closed.

bug #52248